[client/components] Move Post to its own file
- Doing this one by one because it failed earlier
