+++ client/js/app.js
@@ -1,21 +1,11 @@
 require('normalize-css');
 var React = require('react');
 var qwest = require('qwest');
-var marked = require('marked');
-var yaml = require('js-yaml');
-var moment = require('moment');
+
+var Post = require('./Post');
 
 // var url = 'https://api.github.com/repos/spinningarrow/spinningarrow.github.io/contents/_posts';
 var url = 'temp-posts.json';
-
-var parsePost = function (postContent) {
-	var frontMatterEndIndex = postContent.lastIndexOf('---') + 3;
-
-	return {
-		meta: yaml.safeLoad(postContent.substring(3, frontMatterEndIndex - 3)),
-		body: marked(postContent.substring(frontMatterEndIndex))
-	};
-}
 
 var PostBox = React.createClass({
 	getInitialState: function () {
@@ -63,21 +53,6 @@ var PostList = React.createClass({
 
 		return (
 			<ul className="post-list">{nodes}</ul>
-		);
-	}
-});
-
-var Post = React.createClass({
-	getPostMeta: function () {
-		return parsePost(decodeURIComponent(escape(atob(this.props.content)))).meta;
-	},
-
-	render: function () {
-		return (
-			<div className="post">
-				<h2>{this.getPostMeta().title}</h2>
-				<time>{this.getPostMeta().date && moment(this.getPostMeta().date).fromNow()}</time>
-			</div>
 		);
 	}
 });
